rm deprecated  autotransformations
